+++ scrapy/contrib/downloadermiddleware/useragent.py
@@ -6,13 +6,18 @@ from scrapy.utils.python import WeakKeyCache
 class UserAgentMiddleware(object):
     """This middleware allows spiders to override the user_agent"""
 
-    def __init__(self):
+    def __init__(self, user_agent='Scrapy'):
         self.cache = WeakKeyCache(self._user_agent)
+        self.user_agent = user_agent
+
+    @classmethod
+    def from_crawler(cls, crawler):
+        return cls(crawler.settings['USER_AGENT'])
 
     def _user_agent(self, spider):
         if hasattr(spider, 'user_agent'):
             return spider.user_agent
-        return spider.settings['USER_AGENT']
+        return self.user_agent
 
     def process_request(self, request, spider):
         ua = self.cache[spider]