+++ scrapy/contrib/memdebug.py
@@ -10,23 +10,26 @@ from scrapy.xlib.pydispatch import dispatcher
 
 from scrapy import signals
 from scrapy.exceptions import NotConfigured
-from scrapy.conf import settings
 from scrapy.stats import stats
 from scrapy.utils.trackref import live_refs
 
 class MemoryDebugger(object):
 
-    def __init__(self):
+    def __init__(self, trackrefs=False):
         try:
             import libxml2
             self.libxml2 = libxml2
         except ImportError:
             self.libxml2 = None
-        if not settings.getbool('MEMDEBUG_ENABLED'):
-            raise NotConfigured
-
+        self.trackrefs = trackrefs
         dispatcher.connect(self.engine_started, signals.engine_started)
         dispatcher.connect(self.engine_stopped, signals.engine_stopped)
+
+    @classmethod
+    def from_crawler(cls, crawler):
+        if not crawler.settings.getbool('MEMDEBUG_ENABLED'):
+            raise NotConfigured
+        return cls(crawler.settings.getbool('TRACK_REFS'))
 
     def engine_started(self):
         if self.libxml2:
@@ -38,7 +41,7 @@ class MemoryDebugger(object):
             stats.set_value('memdebug/libxml2_leaked_bytes', self.libxml2.debugMemory(1))
         gc.collect()
         stats.set_value('memdebug/gc_garbage_count', len(gc.garbage))
-        if settings.getbool('TRACK_REFS'):
+        if self.trackrefs:
             for cls, wdict in live_refs.iteritems():
                 if not wdict:
                     continue