+++ scrapy/contrib/spiders/crawl.py
@@ -10,7 +10,6 @@ import copy
 from scrapy.http import Request, HtmlResponse
 from scrapy.utils.spider import iterate_spider_output
 from scrapy.spider import BaseSpider
-from scrapy.conf import settings
 
 def identity(x):
     return x
@@ -70,10 +69,9 @@ class CrawlSpider(BaseSpider):
             for requests_or_item in iterate_spider_output(cb_res):
                 yield requests_or_item
 
-        if follow and settings.getbool('CRAWLSPIDER_FOLLOW_LINKS', True):
+        if follow and self._follow_links:
             for request_or_item in self._requests_to_follow(response):
                 yield request_or_item
-                
 
     def _compile_rules(self):
         def get_method(method):
@@ -87,3 +85,7 @@ class CrawlSpider(BaseSpider):
             rule.callback = get_method(rule.callback)
             rule.process_links = get_method(rule.process_links)
             rule.process_request = get_method(rule.process_request)
+
+    def set_crawler(self, crawler):
+        super(CrawlSpider, self).set_crawler(crawler)
+        self._follow_links = crawler.settings.getbool('CRAWLSPIDER_FOLLOW_LINKS', True)