+++ scrapy/tests/test_downloadermiddleware_downloadtimeout.py
@@ -13,7 +13,7 @@ class DownloadTimeoutMiddlewareTest(unittest.TestCase):
         spider = BaseSpider('foo')
         spider.set_crawler(crawler)
         request = Request('http://scrapytest.org/')
-        return request, spider, DownloadTimeoutMiddleware()
+        return request, spider, DownloadTimeoutMiddleware.from_crawler(crawler)
 
     def test_default_download_timeout(self):
         req, spider, mw = self.get_request_spider_mw()
@@ -22,13 +22,13 @@ class DownloadTimeoutMiddlewareTest(unittest.TestCase):
 
     def test_spider_has_download_timeout(self):
         req, spider, mw = self.get_request_spider_mw()
-        spider.DOWNLOAD_TIMEOUT = 2
+        spider.download_timeout = 2
         assert mw.process_request(req, spider) is None
         self.assertEquals(req.meta.get('download_timeout'), 2)
 
     def test_request_has_download_timeout(self):
         req, spider, mw = self.get_request_spider_mw()
-        spider.DOWNLOAD_TIMEOUT = 2
+        spider.download_timeout = 2
         req.meta['download_timeout'] = 1
         assert mw.process_request(req, spider) is None
         self.assertEquals(req.meta.get('download_timeout'), 1)