+++ scrapy/tests/test_settings.py
@@ -1,6 +1,6 @@
 import unittest
 
-from scrapy.settings import Settings, SpiderSettings
+from scrapy.settings import Settings
 from scrapy.utils.test import get_crawler
 from scrapy.spider import BaseSpider
 
@@ -68,35 +68,6 @@ class CrawlerSettingsTest(unittest.TestCase):
         crawler.settings.overrides['DOWNLOAD_TIMEOUT'] = '15'
         self.assertEqual(crawler.settings.getint('DOWNLOAD_TIMEOUT'), 15)
 
-class SpiderSettingsTest(unittest.TestCase):
-
-    def test_global_defaults(self):
-        crawler = get_crawler()
-        settings = SpiderSettings(BaseSpider('name'), crawler.settings)
-        self.assertEqual(settings.getint('DOWNLOAD_TIMEOUT'), 180)
-
-    def test_defaults(self):
-        crawler = get_crawler()
-        crawler.settings.defaults['DOWNLOAD_TIMEOUT'] = '99'
-        settings = SpiderSettings(BaseSpider('name'), crawler.settings)
-        self.assertEqual(settings.getint('DOWNLOAD_TIMEOUT'), 99)
-
-    def test_crawler_defaults(self):
-        crawler = get_crawler({'DOWNLOAD_TIMEOUT': '3'})
-        settings = SpiderSettings(BaseSpider('name'), crawler.settings)
-        self.assertEqual(settings.getint('DOWNLOAD_TIMEOUT'), 3)
-
-    def test_spider_overrides_crawler(self):
-        crawler = get_crawler({'DOWNLOAD_TIMEOUT': '3'})
-        crawler.settings.defaults['DOWNLOAD_TIMEOUT'] = '99'
-        settings = SpiderSettings(BaseSpider('name', DOWNLOAD_TIMEOUT='12'), crawler.settings)
-        self.assertEqual(settings.getint('DOWNLOAD_TIMEOUT'), 12)
-
-    def test_overrides_most_precedence(self):
-        crawler = get_crawler({'DOWNLOAD_TIMEOUT': '3'})
-        crawler.settings.overrides['DOWNLOAD_TIMEOUT'] = '15'
-        settings = SpiderSettings(BaseSpider('name', DOWNLOAD_TIMEOUT='12'), crawler.settings)
-        self.assertEqual(settings.getint('DOWNLOAD_TIMEOUT'), 15)
 
 if __name__ == "__main__":
     unittest.main()